move the HD wrappers for Signer to the signer module
